Add CI to validate nomad jobs
This commit adds a GitHub actions workflow that installs the Nomad CLI and performs validation
on all nomad jobs. This is done automatically via the `validate-jobs.sh` script which finds
all files in the repository with the `.nomad` extension and runs `nomad validate` on them.

This commit also includes enabling dependabot for Terraform.

Signed-off-by: David Bond <davidsbond93@gmail.com>