Add daily job to keep ubuntu up-to-date
This commit adds a sysbatch job that will periodically run updates for ubuntu
packages and the distribution itself. The version of ubuntu on my homelab fell
into EOF which was a massive pain to upgrade involving lots of manual steps. This
job should keep the clients on an LTS version of ubuntu as well as keeping other system
packages up-to-date.

Signed-off-by: David Bond <davidsbond93@gmail.com>