Switch to script for ubuntu upgrades
This commit modifies the ubuntu upgrade job to use a script from this repository
rather than doing all the complexities of scheduling, lifecycle etc. The reason
for switching to this is that I also wanted to perform an `apt-get clean` but couldn't
find a nice way to fit it all in with the lifecycle.

This should also resolve the issues with apt locking as now there's a single task that
performs all the steps one after the other via a script.

Signed-off-by: David Bond <davidsbond93@gmail.com>