Get grafana working with persistent storage
This commit modifies the grafana job to use the root user and mount a CSI
volume at /var/lib/grafana. After trying a multitude of things and googling
a lot the only way I could avoid FS permissions was to use the root user. This
probably isn't the best way but it does work.

Signed-off-by: David Bond <davidsbond93@gmail.com>