Ensure traefik runs on all nodes in cluster
This commit modifies the DNS records to include the nomad server which now
runs as both a server and client. This is not technically recommended but for
my use case isn't particularly worrying.

It allows hitting the endpoints of any of the nomad clients and the filtering
for the 0th node has been removed.

Signed-off-by: David Bond <davidsbond93@gmail.com>