Add speed-dial application (#83)
This commit contains the first homebrewed application for the nomad cluster. It's an application that acts
as a speed-dial for things running in the homelab. The reason for adding this is that I ran out of links
on the chrome new tab page, which caps at 10.

It also includes workflows which should only run when things relating to the application have changed.

Signed-off-by: David Bond <davidsbond93@gmail.com>