Use commits instead of git tags for images (#85)
This commit modifies the goreleaser workflow to use commits for image tag names
rather than git tags. If we have multiple apps in here we don't want the versioning
to be weird where the first version of a new app's image is something like 0.5.0 etc, so
from now on all images just use the commit sha and we can make tag names anything we
want to trigger the workflow.

The vet job should also run on pushes to master now.

Signed-off-by: David Bond <davidsbond93@gmail.com>