Add command for listing container images
This commit adds the `homadctl containers list` command that will return a list
of all container images in use in the nomad cluster. Specifically it looks for jobs
using the docker driver and grabs the image from their configuration.

This is a precursor to creating a command to check for available image upgrades.

Signed-off-by: David Bond <davidsbond93@gmail.com>